Enterprise-Knowledge-Assistant/services/fastapi-query/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pinecone-client==3.2.2
openai==1.58.0
httpx==0.27.0
prometheus-client==0.20.0
pydantic==2.7.0